% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.prime.R
\name{is.prime}
\alias{is.prime}
\title{Testing if Prime}
\usage{
is.prime(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
boolean vector indicating prime numbers
}
\description{
The function takes a vector of positive integers greater than or equal to 2
and returns a boolean vector indicating whether it is prime or not.
}
\examples{
x <- c(2,3,4)
is.prime(x)
}
